#774 Move the SWG back-pressure stress test into the engine suite, where engine changes run it
The test asserts an engine property — that the commit path sustains 105,500 spawns across a five-component indexed schema without the page cache raising back-pressure — but it lived in Typhon.Workbench.Tests purely because the fixture generator it drives is a Workbench tool. The merge gate runs that project only when the path filter tools/Typhon.Workbench/** matches, so a change confined to src/Typhon.Engine/** never executed the gate's heaviest back-pressure case. Confirmed across three consecutive merges: the runs for #765 and #769 did not run it at all, and #772's run did only because that branch happened to touch one SPA file. The gate was reporting green on engine merges while its hardest engine case sat unrun.

Quarantined against #774 on arrival rather than left blocking. It fails 2 runs out of 2 on the self-hosted gate box with "B+Tree insert made no progress in 10000 pessimistic retries" and passes locally in Release, so it is reliably red rather than flaky — the second run's engine shard was initial-failed=0, meaning no retry net could absorb it, and by shard.py's own premise that a real regression fails every attempt, none should be asked to. The nightly still runs it, so the quarantine records the state instead of hiding it.

The exception's own "(see #695)" citation is stale: #695 was the unbounded SpinWriteLock livelock and is closed, fixed by bounding the spin and throwing. What the throw now reports is the #739 family, which is still open.
