+++ test/Typhon.Engine.Tests/QUARANTINE.md
@@ -68,4 +68,5 @@ The canonical red set is **platform-specific** and must be determined on the CI 
 | `Typhon.Engine.Tests.ChaosStressTests` — 7 tests (`CrossEntityTransaction_AtomicMultiUpdate`, `AllowMultipleIndex_HighChurn`, `IndexSplit_CascadingSplitsUnderContention`, `RollbackStress_ConcurrentRollbacks`, `UniqueIndexViolation_UnderLoad`, `UltimateStress_AllSubsystems`) | [#696](https://github.com/log2n-io/Typhon/issues/696) | Concurrent value/index correctness failures, 3/3 on first-ever execution. The fixture as a whole is now `[Explicit] + [Category("Nightly")]`; these individual tests are additionally quarantined. | 2026-08-07 |
 | `Typhon.Engine.Tests.ChaosStressTests.CreateDeleteRecreate_RapidLifecycle` | [#696](https://github.com/log2n-io/Typhon/issues/696) | **Retargeted from #695, whose livelock is fixed.** It used to need `--blame-hang` to terminate; it now completes in ~150 ms and fails about 1 run in 4 with *"A duplicate key was detected in a unique index"* — a concurrent unique-index correctness defect, #696's family, possibly #716's mechanism. | 2026-08-08 |
 | `Typhon.Engine.Tests.LinearizabilityTests.ParallelOperations_AreLinearizable` (+ `_Deep`, + `DirtyCounter_IsConservedAtQuiesce`) | [#400](https://github.com/log2n-io/Typhon/issues/400) | The T5 model, run on the **production default** (`Deferred` + parallel transactions on one shared `UnitOfWork`), reproduces the shared-`ChangeSet` race within milliseconds: the new `[Conditional("DEBUG")]` concurrent-mutation guard reports e.g. *"thread 27 entered AddByMemPageIndex while thread 23 was still inside a mutating method"*. That is the T5 result, not an obstacle to it. Forcing these green — by moving to `Immediate`, or by removing the guard — would restore exactly the 188:48 selection bias the fixture exists to remove. Note the guard reddens **nothing else** in 5,003 tests, which measures how completely the racy structure was going unexercised. | 2026-08-07 |
+| `Typhon.Engine.Tests.Data.SwgFixtureStressTests.Stress_Config_Completes_Without_Backpressure_Timeout` | [#774](https://github.com/log2n-io/Typhon/issues/774) | **Moved here from `Typhon.Workbench.Tests.FixtureConfigTests`, where the gate's path filter had been hiding it.** Reliably red, not flaky: 2 runs of 2 on the self-hosted gate box with *"B+Tree insert made no progress in 10000 pessimistic retries"*, green locally in Release. The second run's engine shard was `initial-failed=0`, so a retry net cannot absorb it — and by `shard.py`'s own premise ("a REAL regression fails every attempt") it must not be asked to. The underlying liveness defect is the #739 family; the exception's `(see #695)` citation is stale, #695's livelock being fixed. | 2026-08-11 |
 | `Typhon.Engine.Tests.DifferentialRecoveryOracleTests.PayloadAxes_SurviveACrash` | [#389](https://github.com/log2n-io/Typhon/issues/389) | `ComponentCollection` buffer mutations are not WAL-redo-logged, so after a crash in the WAL window every collection recovers **EMPTY behind an intact buffer descriptor**. This is the first test in the suite that can SEE that: the raw component-bytes comparison reports 0 mismatches (the descriptor survives) while the element check reports all 8 — which is precisely why `RecoveryShadowModel` now refuses to capture a collection-bearing archetype without an `ICollectionProjector`. Regression lock for when #389 is fixed. | 2026-08-07 |
